refactor:에러 핸들러 위치 변경, json 응답 통일, UUID생성 함수 수정
에러 핸들러의 위치를 services 내에서 src로 이동합니다
요청/응답 모두 json 결과를 전달하도록 수정합니다
UUID를 생성하는 함수가 Crypto 모듈을 사용해서 랜덤생성하도록 수정합니다
