+++ back/src/api/user/balance.ts
@@ -1,5 +1,5 @@
 import express, { NextFunction, Request, Response } from 'express';
-import { AuthError, AuthErrorMessage, ParamError, ParamErrorMessage } from '@services/errors/index';
+import { AuthError, AuthErrorMessage, ParamError, ParamErrorMessage } from 'errors/index';
 import { UserService } from '@services/index';
 import { IBalanceHistory, BALANCETYPE, STATUSTYPE } from '@models/index';
