+++ back/src/errors/AuthError.ts
