+++ back/src/errors/CommonError.ts
