+++ back/tsconfig.json
@@ -40,7 +40,8 @@
       "@repositories/*": ["repositories/*"],
       "@services/*": ["services/*"],
       "@types/*": ["types/*"],
-      "@loaders/*": ["loaders/*"]
+      "@loaders/*": ["loaders/*"],
+      "@errors/*": ["errors/*"],
     }, /* Specify a set of entries that re-map imports to additional lookup locations. */
     // "rootDirs": [],                                   /* Allow multiple folders to be treated as one when resolving modules. */
     "typeRoots": [
